modified jwt provider , created separate file
