+++ src/main/resources/application.properties
@@ -1,6 +1,10 @@
 spring.main.banner-mode = console
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=jdbc:mysql://localhost:3306/user_auth
 spring.datasource.username=root
 spring.datasource.password=root
 spring.jpa.show-sql=true
+
+#Jwt
+secret = user-auth
+tokenValidity = 100000000
